remove ContactDetailCoordinator
